# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restau TyTy API",
        "description": "API documentation for Restaurant Management System",
        "contact": {
            "email": "admin@restautyty.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://restau-bo.test",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/commandes/{commandeId}/paiements": {
            "get": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Get list of paiements",
                "operationId": "getPaiementsList",
                "parameters": [
                    {
                        "name": "commandeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Store new paiement",
                "description": "Guest requires X-Order-Token",
                "operationId": "storePaiement",
                "parameters": [
                    {
                        "name": "commandeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Order-Token",
                        "in": "header",
                        "description": "Guest Session Token",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Paiement"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/v1/commandes/{commandeId}/paiements/{id}": {
            "get": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Get paiement information",
                "operationId": "getPaiementById",
                "parameters": [
                    {
                        "name": "commandeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Update paiement",
                "operationId": "updatePaiement",
                "parameters": [
                    {
                        "name": "commandeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Paiement"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Delete paiement",
                "operationId": "deletePaiement",
                "parameters": [
                    {
                        "name": "commandeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/categories/{categorieId}/plats": {
            "get": {
                "tags": [
                    "Plats"
                ],
                "summary": "Get list of plats",
                "operationId": "getPlatsList",
                "parameters": [
                    {
                        "name": "categorieId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Plats"
                ],
                "summary": "Store new plat",
                "operationId": "storePlat",
                "parameters": [
                    {
                        "name": "categorieId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Plat"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/categories/{categorieId}/plats/{id}": {
            "get": {
                "tags": [
                    "Plats"
                ],
                "summary": "Get plat information",
                "operationId": "getPlatById",
                "parameters": [
                    {
                        "name": "categorieId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Plats"
                ],
                "summary": "Update plat",
                "operationId": "updatePlat",
                "parameters": [
                    {
                        "name": "categorieId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Plat"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Plats"
                ],
                "summary": "Delete plat",
                "operationId": "deletePlat",
                "parameters": [
                    {
                        "name": "categorieId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Paiement": {
                "required": [
                    "commande_id",
                    "montant",
                    "mode",
                    "statut"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "commande_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "montant": {
                        "type": "number",
                        "format": "float",
                        "example": 25
                    },
                    "mode": {
                        "type": "string",
                        "example": "cash"
                    },
                    "statut": {
                        "type": "string",
                        "example": "completed"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Plat": {
                "required": [
                    "nom",
                    "prix",
                    "categorie_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "Burger Classique"
                    },
                    "description": {
                        "type": "string",
                        "example": "Boeuf, cheddar, salade, tomate"
                    },
                    "prix": {
                        "type": "number",
                        "format": "float",
                        "example": 12.5
                    },
                    "image": {
                        "type": "string",
                        "nullable": true
                    },
                    "categorie_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}